backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
sqlalchemy==1.4.53
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pdfplumber==0.10.3
